removed gtm code, retained event types
