+++ package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@next/third-parties": "^14.2.10",
     "geist": "^1.3.1",
     "next": "14.2.6",
     "react": "^18",
