+++ src/lib/modules/temp/GoogleTagManagerTypes.ts
@@ -1,0 +1,10 @@
+interface SystemInformation {
+    browser_type: string,
+    device_type: string,
+    os_type: string,
+    screen_resolution: string,
+}
+
+export type {
+    SystemInformation
+}
