+++ src/lib/modules/temp/Interactions.ts
@@ -1,4 +1,4 @@
-export enum InteractionType {
+export enum MouseInteraction {
     CLICK = "Click",
     MOUSE_OVER = "MouseOver",
     MOUSE_OUT = "MouseOut",
